Move factory method that creates `Property` instances from `DefaultProviderFactory` to `DefaultObjectFactory`.
Change `DefaultSourceDirectorySet` to use `ObjectFactory` to create property instances, rather than reaching in to some internal service.
