+++ subprojects/core/src/main/java/org/gradle/api/internal/model/DefaultObjectFactory.java
@@ -18,8 +18,9 @@ package org.gradle.api.internal.model;
 
 import org.gradle.api.Named;
 import org.gradle.api.internal.provider.DefaultListProperty;
-import org.gradle.api.internal.provider.DefaultProviderFactory;
+import org.gradle.api.internal.provider.DefaultPropertyState;
 import org.gradle.api.internal.provider.DefaultSetProperty;
+import org.gradle.api.internal.provider.Providers;
 import org.gradle.api.model.ObjectFactory;
 import org.gradle.api.provider.ListProperty;
 import org.gradle.api.provider.Property;
@@ -30,12 +31,10 @@ import org.gradle.internal.reflect.Instantiator;
 public class DefaultObjectFactory implements ObjectFactory {
     private final Instantiator instantiator;
     private final NamedObjectInstantiator namedObjectInstantiator;
-    private final DefaultProviderFactory providerFactory;
 
     public DefaultObjectFactory(Instantiator instantiator, NamedObjectInstantiator namedObjectInstantiator) {
         this.instantiator = instantiator;
         this.namedObjectInstantiator = namedObjectInstantiator;
-        providerFactory = new DefaultProviderFactory();
     }
 
     @Override
@@ -50,7 +49,31 @@ public class DefaultObjectFactory implements ObjectFactory {
 
     @Override
     public <T> Property<T> property(Class<T> valueType) {
-        return providerFactory.propertyNoNag(valueType);
+        if (valueType == null) {
+            throw new IllegalArgumentException("Class cannot be null");
+        }
+
+        Property<T> property = new DefaultPropertyState<T>(valueType);
+
+        if (valueType == Boolean.class) {
+            ((Property<Boolean>) property).set(Providers.FALSE);
+        } else if (valueType == Byte.class) {
+            ((Property<Byte>) property).set(Providers.BYTE_ZERO);
+        } else if (valueType == Short.class) {
+            ((Property<Short>) property).set(Providers.SHORT_ZERO);
+        } else if (valueType == Integer.class) {
+            ((Property<Integer>) property).set(Providers.INTEGER_ZERO);
+        } else if (valueType == Long.class) {
+            ((Property<Long>) property).set(Providers.LONG_ZERO);
+        } else if (valueType == Float.class) {
+            ((Property<Float>) property).set(Providers.FLOAT_ZERO);
+        } else if (valueType == Double.class) {
+            ((Property<Double>) property).set(Providers.DOUBLE_ZERO);
+        } else if (valueType == Character.class) {
+            ((Property<Character>) property).set(Providers.CHAR_ZERO);
+        }
+
+        return property;
     }
 
     @Override
