+++ subprojects/core/src/main/java/org/gradle/internal/service/scopes/ProjectScopeServices.java
@@ -134,8 +134,8 @@ public class ProjectScopeServices extends DefaultServiceRegistry {
         return new BaseDirFileResolver(get(FileSystem.class), project.getProjectDir(), patternSetFactory);
     }
 
-    protected SourceDirectorySetFactory createSourceDirectorySetFactory(FileResolver fileResolver, DirectoryFileTreeFactory directoryFileTreeFactory) {
-        return new DefaultSourceDirectorySetFactory(fileResolver, directoryFileTreeFactory);
+    protected SourceDirectorySetFactory createSourceDirectorySetFactory(FileResolver fileResolver, DirectoryFileTreeFactory directoryFileTreeFactory, ObjectFactory objectFactory) {
+        return new DefaultSourceDirectorySetFactory(fileResolver, directoryFileTreeFactory, objectFactory);
     }
 
     protected LoggingManagerInternal createLoggingManager() {
