+++ subprojects/core/src/test/groovy/org/gradle/api/internal/file/DefaultSourceDirectorySetTest.groovy
@@ -23,8 +23,10 @@ import org.gradle.api.file.FileTree
 import org.gradle.api.file.SourceDirectorySet
 import org.gradle.api.internal.file.collections.DefaultConfigurableFileCollection
 import org.gradle.api.internal.file.collections.DirectoryFileTreeFactory
+import org.gradle.api.model.ObjectFactory
 import org.gradle.test.fixtures.file.TestFile
 import org.gradle.test.fixtures.file.TestNameTestDirectoryProvider
+import org.gradle.util.TestUtil
 import org.gradle.util.UsesNativeServices
 import org.junit.Rule
 import spock.lang.Specification
@@ -39,10 +41,11 @@ class DefaultSourceDirectorySetTest extends Specification {
     private final TestFile testDir = tmpDir.testDirectory
     private FileResolver resolver = TestFiles.resolver(testDir)
     private DirectoryFileTreeFactory directoryFileTreeFactory = TestFiles.directoryFileTreeFactory()
+    private ObjectFactory objectFactory = TestUtil.objectFactory()
     private DefaultSourceDirectorySet set
 
     void setup() {
-        set = new DefaultSourceDirectorySet('<display-name>', resolver, directoryFileTreeFactory)
+        set = new DefaultSourceDirectorySet('files', '<display-name>', resolver, directoryFileTreeFactory, objectFactory)
     }
 
     void hasUsefulToString() {
@@ -84,7 +87,7 @@ class DefaultSourceDirectorySetTest extends Specification {
     }
 
     void addsContentsOfAnotherSourceDirectorySet() {
-        SourceDirectorySet nested = new DefaultSourceDirectorySet('<nested>', resolver, directoryFileTreeFactory)
+        SourceDirectorySet nested = new DefaultSourceDirectorySet('nested', '<nested>', resolver, directoryFileTreeFactory, objectFactory)
         nested.srcDir 'dir1'
 
         when:
@@ -101,7 +104,7 @@ class DefaultSourceDirectorySetTest extends Specification {
     }
 
     void addsSourceDirectoriesOfAnotherSourceDirectorySet() {
-        SourceDirectorySet nested = new DefaultSourceDirectorySet('<nested>', resolver, directoryFileTreeFactory)
+        SourceDirectorySet nested = new DefaultSourceDirectorySet('nested', '<nested>', resolver, directoryFileTreeFactory, objectFactory)
         nested.srcDir 'dir1'
 
         when:
@@ -118,7 +121,7 @@ class DefaultSourceDirectorySetTest extends Specification {
     }
 
     void settingSourceDirsReplacesExistingContent() {
-        SourceDirectorySet nested = new DefaultSourceDirectorySet('<nested>', resolver, directoryFileTreeFactory)
+        SourceDirectorySet nested = new DefaultSourceDirectorySet('nested', '<nested>', resolver, directoryFileTreeFactory, objectFactory)
         nested.srcDir 'ignore me'
         set.srcDir 'ignore me as well'
         set.source nested
@@ -185,7 +188,7 @@ class DefaultSourceDirectorySetTest extends Specification {
     }
 
     void convertsNestedDirectorySetsToDirectoryTrees() {
-        SourceDirectorySet nested = new DefaultSourceDirectorySet('<nested>', resolver, directoryFileTreeFactory)
+        SourceDirectorySet nested = new DefaultSourceDirectorySet('nested', '<nested>', resolver, directoryFileTreeFactory, objectFactory)
         nested.srcDirs 'dir1', 'dir2'
 
         when:
@@ -199,7 +202,7 @@ class DefaultSourceDirectorySetTest extends Specification {
     }
 
     void removesDuplicateDirectoryTrees() {
-        SourceDirectorySet nested = new DefaultSourceDirectorySet('<nested>', resolver, directoryFileTreeFactory)
+        SourceDirectorySet nested = new DefaultSourceDirectorySet('nested', '<nested>', resolver, directoryFileTreeFactory, objectFactory)
         nested.srcDirs 'dir1', 'dir2'
 
         when:
@@ -306,8 +309,8 @@ class DefaultSourceDirectorySetTest extends Specification {
 
     void setAndItsViewsHaveDependenciesOfAllSourceDirectorySets() {
         given:
-        def nested1 = new DefaultSourceDirectorySet('<nested-1>', resolver, directoryFileTreeFactory)
-        def nested2 = new DefaultSourceDirectorySet('<nested-2>', resolver, directoryFileTreeFactory)
+        def nested1 = new DefaultSourceDirectorySet('nested-1', '<nested-1>', resolver, directoryFileTreeFactory, objectFactory)
+        def nested2 = new DefaultSourceDirectorySet('nested-2', '<nested-2>', resolver, directoryFileTreeFactory, objectFactory)
         def task1 = Stub(Task)
         def task2 = Stub(Task)
         nested1.srcDir dir("dir1", task1)
