+++ subprojects/core/src/test/groovy/org/gradle/api/internal/provider/DefaultProviderFactoryTest.groovy
@@ -58,67 +58,4 @@ class DefaultProviderFactoryTest extends Specification {
         String    | 'hello'
         File      | TEST_FILE
     }
-
-    def "cannot create property for null value"() {
-        when:
-        providerFactory.propertyNoNag(null)
-
-        then:
-        def t = thrown(IllegalArgumentException)
-        t.message == 'Class cannot be null'
-    }
-
-    @Unroll
-    def "property state representing boolean and numbers provide default value for #type"() {
-        given:
-        def property = providerFactory.propertyNoNag(type)
-
-        expect:
-        property.get() == defaultValue
-
-        where:
-        type      | defaultValue
-        Boolean   | false
-        Byte      | 0
-        Short     | 0
-        Integer   | 0
-        Long      | 0L
-        Float     | 0.0f
-        Double    | 0.0d
-        Character | '\u0000'
-    }
-
-    @Unroll
-    def "can create property state for #type"() {
-        when:
-        def property = providerFactory.propertyNoNag(type)
-        property.set(value)
-
-        then:
-        property
-        property.get() == value
-
-        where:
-        type      | value
-        Boolean   | true
-        Byte      | (byte) 1
-        Short     | (short) 2
-        Integer   | 3
-        Long      | 4L
-        Float     | 5.5f
-        Double    | 6.6d
-        Character | (char) '\u1234'
-        String    | 'hello'
-        File      | TEST_FILE
-    }
-
-    def "creating property type for reference type throws exception upon retrieval of value"() {
-        when:
-        def property = providerFactory.propertyNoNag(Runnable)
-        property.get()
-
-        then:
-        def t = thrown(IllegalStateException)
-        t.message == 'No value has been specified for this provider.'
-    }
 }
