+++ subprojects/core/src/testFixtures/groovy/org/gradle/api/internal/file/TestFiles.java
@@ -36,6 +36,7 @@ import org.gradle.process.internal.ExecFactory;
 import org.gradle.process.internal.ExecHandleFactory;
 import org.gradle.process.internal.JavaExecHandleFactory;
 import org.gradle.testfixtures.internal.NativeServicesTestFixture;
+import org.gradle.util.TestUtil;
 
 import java.io.File;
 
@@ -109,11 +110,11 @@ public class TestFiles {
     }
 
     public static SourceDirectorySetFactory sourceDirectorySetFactory() {
-        return new DefaultSourceDirectorySetFactory(resolver(), new DefaultDirectoryFileTreeFactory());
+        return new DefaultSourceDirectorySetFactory(resolver(), new DefaultDirectoryFileTreeFactory(), TestUtil.objectFactory());
     }
 
     public static SourceDirectorySetFactory sourceDirectorySetFactory(File baseDir) {
-        return new DefaultSourceDirectorySetFactory(resolver(baseDir), new DefaultDirectoryFileTreeFactory());
+        return new DefaultSourceDirectorySetFactory(resolver(baseDir), new DefaultDirectoryFileTreeFactory(), TestUtil.objectFactory());
     }
 
     public static ExecFactory execFactory() {
